LayoutTests/platform/qt/fast/files/url-required-arguments-expected.txt

Test required arguments of URL object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS webkitURL.createObjectURL() threw exception TypeError: Not enough arguments.
PASS webkitURL.revokeObjectURL() threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
